feat: Rename Status queries to State
