+++ instances/instance_state.go
@@ -3,7 +3,7 @@
 // Licensed under the BSD-3-Clause License (the "License").
 // You may not use this file except in compliance with the License.
 
-package services
+package instances
 
 import (
 	"context"
@@ -14,18 +14,25 @@ import (
 	"sdk.kraft.cloud/client"
 )
 
-// Status returns the current status and the configuration of a service group.
-func (c *servicesClient) Status(ctx context.Context, uuidOrName string) (*ServiceGroup, error) {
+// State returns the current state and the configuration of an instance.
+// UUIDs can also be names.
+//
+// See: https://docs.kraft.cloud/002-rest-api-v1-instances.html#state
+func (c *instancesClient) State(ctx context.Context, uuidOrName string) (*Instance, error) {
 	if uuidOrName == "" {
 		return nil, errors.New("UUID cannot be empty")
 	}
 
 	endpoint := Endpoint + "/" + uuidOrName
 
-	var response client.ServiceResponse[ServiceGroup]
+	var response client.ServiceResponse[Instance]
 	if err := c.request.DoRequest(ctx, http.MethodGet, endpoint, nil, &response); err != nil {
 		return nil, fmt.Errorf("performing the request: %w", err)
 	}
 
-	return response.FirstOrErr()
+	instance, err := response.FirstOrErr()
+	if instance != nil && instance.Message != "" {
+		err = fmt.Errorf("%w: %s", err, instance.Message)
+	}
+	return instance, err
 }
