+++ services/service_state.go
@@ -3,7 +3,7 @@
 // Licensed under the BSD-3-Clause License (the "License").
 // You may not use this file except in compliance with the License.
 
-package instances
+package services
 
 import (
 	"context"
@@ -14,24 +14,19 @@ import (
 	"sdk.kraft.cloud/client"
 )
 
-// Status returns the current status and the configuration of an instance.
-//
-// See: https://docs.kraft.cloud/002-rest-api-v1-instances.html#status
-func (c *instancesClient) Status(ctx context.Context, uuidOrName string) (*Instance, error) {
+// State returns the current state and the configuration of a service group.
+// UUIDs can also be names.
+func (c *servicesClient) State(ctx context.Context, uuidOrName string) (*ServiceGroup, error) {
 	if uuidOrName == "" {
 		return nil, errors.New("UUID cannot be empty")
 	}
 
 	endpoint := Endpoint + "/" + uuidOrName
 
-	var response client.ServiceResponse[Instance]
+	var response client.ServiceResponse[ServiceGroup]
 	if err := c.request.DoRequest(ctx, http.MethodGet, endpoint, nil, &response); err != nil {
 		return nil, fmt.Errorf("performing the request: %w", err)
 	}
 
-	instance, err := response.FirstOrErr()
-	if instance != nil && instance.Message != "" {
-		err = fmt.Errorf("%w: %s", err, instance.Message)
-	}
-	return instance, err
+	return response.FirstOrErr()
 }
